Milk frothing minigame
You can now drag the jug up and down. There is an indicator on the left side which will turn green and yellow based on how close you are to the target slider when dragging the jug. If you get too far it wil turn red and spoil the milk.