Latte art mechanic base complete
The minigame will now take 3 inputs from WASD which need to be specific combinations. Getting the combinations correct will give you the relative latte art. If you do a design that is not correct it will throw a failed design. There's a button to show the latte art menu aswell.